pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "chaos-runner"
version = "1.0.0"
description = "AI-Powered Resilience Gate for CI/CD Pipelines — ChaosController CLI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "ChaosController Project"}
]
keywords = ["chaos-engineering", "resilience", "cicd", "sre", "testing"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Testing",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "click>=8.1",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "pydantic>=2.0",
  "google-genai>=1.0",
  "python-dotenv>=1.0",
]

[project.scripts]
chaos-runner = "chaos_runner.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["chaos_runner*"]
